STRING TRANSFORMERS PROJECT EXTRACTED FROM RECONCILIATION AND MATCHING FRAMEWORK.
Git history prior to this commit may be inconsistent: the code may not
compile or work as expected.  Some commits may make no sense, appear to
make no relevant change, or refer to changes not made.

The String Transformers began as part of a larger piece of software, the
Reconciliation and Matching Framework.  Commits affecting relevant classes
have been extracted to produce this repository, which we hope is useful
on its own.

(This particular commit:
• Converts project to single-project Maven build,
• Renames packages,
• Removes dependency on CSV library, leaving only an interface for the Dictionary,
• Updates to Apache Commons-Lang 3.)